Extracted create account helpers to utils file
