+++ packages/commands/src/accounts/create.ts
@@ -1,24 +1,11 @@
-import confirm from "@inquirer/confirm";
-import { Keypair } from "@solana/web3.js";
 import { Command, Option } from "commander";
-import { ethers } from "ethers";
 import { z } from "zod";
 
 import {
-  AccountData,
   accountNameSchema,
   AvailableAccountTypes,
 } from "../../../../types/accounts.types";
-import {
-  safeExists,
-  safeMkdir,
-  safeWriteFile,
-} from "../../../../utils/fsUtils";
-import { handleError } from "../../../../utils/handleError";
-import {
-  getAccountKeyPath,
-  getAccountTypeDir,
-} from "../../../../utils/keyPaths";
+import { createAccountForType } from "../../../../utils/accounts";
 import { validateAndParseSchema } from "../../../../utils/validateAndParseSchema";
 
 const createAccountOptionsSchema = z.object({
@@ -31,63 +18,6 @@ const createAccountOptionsSchema = z.object({
 });
 
 type CreateAccountOptions = z.infer<typeof createAccountOptionsSchema>;
-
-const createEVMAccount = (): AccountData => {
-  const wallet = ethers.Wallet.createRandom();
-  return {
-    address: wallet.address,
-    mnemonic: wallet.mnemonic?.phrase,
-    privateKey: wallet.privateKey,
-  };
-};
-
-const createSolanaAccount = (): AccountData => {
-  const keypair = Keypair.generate();
-  return {
-    publicKey: keypair.publicKey.toBase58(),
-    secretKey: Buffer.from(keypair.secretKey).toString("hex"),
-  };
-};
-
-const createAccountForType = async (
-  type: (typeof AvailableAccountTypes)[number],
-  name: string
-): Promise<void> => {
-  try {
-    const baseDir = getAccountTypeDir(type);
-    safeMkdir(baseDir);
-
-    const keyPath = getAccountKeyPath(type, name);
-
-    if (safeExists(keyPath)) {
-      const shouldOverwrite = await confirm({
-        default: false,
-        message: `File ${keyPath} already exists. Overwrite?`,
-      });
-      if (!shouldOverwrite) {
-        console.log(`Operation cancelled for ${type} account.`);
-        return;
-      }
-    }
-
-    const keyData = type === "evm" ? createEVMAccount() : createSolanaAccount();
-
-    safeWriteFile(keyPath, keyData);
-    console.log(`${type.toUpperCase()} account created successfully!`);
-    console.log(`Key saved to: ${keyPath}`);
-    if (type === "evm") {
-      console.log(`Address: ${keyData.address}`);
-    } else {
-      console.log(`Public Key: ${keyData.publicKey}`);
-    }
-  } catch (error: unknown) {
-    handleError({
-      context: "Failed to create or save account",
-      error,
-      shouldThrow: true,
-    });
-  }
-};
 
 const main = async (options: CreateAccountOptions) => {
   const { type, name } = options;
