+++ utils/accounts.ts
@@ -1,11 +1,17 @@
+import confirm from "@inquirer/confirm";
+import { Keypair } from "@solana/web3.js";
+import { ethers } from "ethers";
+
 import {
+  AccountData,
   accountDataSchema,
   AvailableAccountTypes,
   EVMAccountData,
   SolanaAccountData,
 } from "../types/accounts.types";
-import { safeExists, safeReadFile } from "./fsUtils";
-import { getAccountKeyPath } from "./keyPaths";
+import { safeExists, safeMkdir, safeReadFile, safeWriteFile } from "./fsUtils";
+import { handleError } from "./handleError";
+import { getAccountKeyPath, getAccountTypeDir } from "./keyPaths";
 import { parseJson } from "./parseJson";
 
 /**
@@ -39,5 +45,62 @@ export const getAccountData = <T extends EVMAccountData | SolanaAccountData>(
     return keyData as T;
   } catch {
     return undefined;
+  }
+};
+
+const createEVMAccount = (): AccountData => {
+  const wallet = ethers.Wallet.createRandom();
+  return {
+    address: wallet.address,
+    mnemonic: wallet.mnemonic?.phrase,
+    privateKey: wallet.privateKey,
+  };
+};
+
+const createSolanaAccount = (): AccountData => {
+  const keypair = Keypair.generate();
+  return {
+    publicKey: keypair.publicKey.toBase58(),
+    secretKey: Buffer.from(keypair.secretKey).toString("hex"),
+  };
+};
+
+export const createAccountForType = async (
+  type: (typeof AvailableAccountTypes)[number],
+  name: string
+): Promise<void> => {
+  try {
+    const baseDir = getAccountTypeDir(type);
+    safeMkdir(baseDir);
+
+    const keyPath = getAccountKeyPath(type, name);
+
+    if (safeExists(keyPath)) {
+      const shouldOverwrite = await confirm({
+        default: false,
+        message: `File ${keyPath} already exists. Overwrite?`,
+      });
+      if (!shouldOverwrite) {
+        console.log(`Operation cancelled for ${type} account.`);
+        return;
+      }
+    }
+
+    const keyData = type === "evm" ? createEVMAccount() : createSolanaAccount();
+
+    safeWriteFile(keyPath, keyData);
+    console.log(`${type.toUpperCase()} account created successfully!`);
+    console.log(`Key saved to: ${keyPath}`);
+    if (type === "evm") {
+      console.log(`Address: ${keyData.address}`);
+    } else {
+      console.log(`Public Key: ${keyData.publicKey}`);
+    }
+  } catch (error: unknown) {
+    handleError({
+      context: "Failed to create or save account",
+      error,
+      shouldThrow: true,
+    });
   }
 };
